Further suggestion that the CODE definition copied from RFC 5531 should be removed from this document.
Magnus says:

"In this particular case it might not be a major issue, but it is
recommended that not put the numbers from the IANA registry into
to the document until it has been assigned to the document.
In cases where one need a number for testing during development,
one can in many registries actually issue an early assignment.

To me the right way of doing this table would have been to replace
the 7 on the AUTH_TLS line with a [IANA_TBA]. And add an RFC-editor
note in this section, and the necessary IANA sub-section requesting
assignment. And if an number would have been needed for the
implementation an early assignment request could have been done,"

Dave notes:

"This was needed but wasn't done then."

Magnus continues:

"and then converted to a permanent one using the IANA section."