Address Gen-ART review comments from Dale Worley:
4.1.  Discovering Server-side TLS Support

Somewhere in this section you need to specify the semi-obvious:

  In principle, RPC is transport-agnostic.  In the context of
  RPC-Over-TLS, the initial request MUST be sent using either TCP or
  UDP.  If an initial TCP request is successful, a TLS connection is
  established.  If an initial UDP request is successful, a DTLS
  association is established.

The point is that if you upgrade, if you start with TCP, you MUST
upgrade to TLS, and if you start with UDP, you MUST upgrade to DTLS.
Whereas it is conceivable that one could start with UDP to port 111,
discover rpc-tls support and then do a TLS connection to TCP port 111
("the same port") to continue.  (After all, every NFS server listens on
111 both with UDP and TCP, right?)  And you have to state that
explicitly as a requirement.

--

  If the Reply's reply_stat is MSG_ACCEPTED but the verifier does not
  contain the "STARTTLS" token, or if the Reply's reply_stat is
  MSG_DENIED, the RPC client MUST NOT send a "ClientHello" message.

Strictly, this should list the three tests required by the preceding
paragraph:

  If the Reply's reply_stat is not MSG_ACCEPTED, if its verifier is
  not an AUTH_NONE, or if its verifier's contents are not "STARTTLS",
  the RPC client MUST NOT send a "ClientHello" message.