Address Gen-ART review comment from Dale Worley
5.2.2.  X.509 Certificates Using Fingerprints

  Implementations MUST support SHA-256
  [FIPS.180-4] or stronger as the hash algorithm for the fingerprint.

Perhaps there is a usage in security, but it seems to me that
"stronger" is not well-defined.  I expect that it means that the hash
algorithm must produce output of at least 256 bits, but that criterion
makes no allowances that over time algorithms might be deprecated.  Is
there a list of accepted "strong" hash algorithms that can be
referenced here to make this unambiguous and track best practice over
time?

Also -- here I'm assuming that one peer presents a certificate to the
other, and the second peer first hashes the certificate to create a
fingerprint and then checks the fingerprint against a list -- this
sentence does not require the peer to *use* an algorithm that meets
this criterion.  The text needs to be something like

  Implementations MUST use SHA-256 [FIPS.180-4] or stronger as the
  hash algorithm for the fingerprint.

----

My choice is to withdraw the SHA-256 requirement:

* There is no standard definition for computing certificate
  fingerprints.
* Most implementations choose SHA-1 already, and I have found no
  authority that claims SHA-1 is inadequate for this purpose.
* Certificate fingerprints are constructed by system components other
  than the RPC implementation.