Address Gen-ART review comment from Dale Worley
7.4.  Best Security Policy Practices

  o  When using RPCSEC_GSS, GSS/Kerberos provides adequate host
     authentication and a policy that requires GSS mutual
     authentication and rejection of a connection when host
     authentication fails.  GSS integrity and privacy services,
     therefore, can be disabled in favor of TLS encryption with peer
     authentication.

This paragraph is not constructed correctly.  The first sentence says
that GSS provides certain facilities, and the second sentence says
that "therefore" GSS (or certain parts of GSS) can be disabled.  There
are a number of possibilities for the form the argument should take,
but I don't know enough about security to know what was intended.