Address IESG ballot COMMENT from Martin Duke
"Moreover, the use of AUTH_SYS remains common despite the adverse effects that
acceptance of UIDs and GIDs from unauthenticated clients brings with it.
Continued use is in part because:

Per-client deployment and administrative costs are not scalable. Administrators
must provide keying material for each RPC client, including transient clients.
Host identity management and user identity management must be enforced in the
same security realm. In certain environments, different authorities might be
responsible for provisioning client systems versus provisioning new users."

The text above says that something is still in use because..., and then
mentions two things that are bad. I think the two items are supposed to refer
to GSS?