Address IESG ballot COMMENT from Martin Duke
Sec 4.2 This sure looks like normative text, so s/must not/MUST NOT "...
utilize the remote TLS peer identity for RPC user authentication"

[ In response, I noted that:

This paragraph formerly had MUST NOT, but was rewritten to address this issue:

 https://github.com/chucklever/i-d-rpc-tls/issues/5

There was a concern that the existing statement was not specific enough to make
it a compliance requirement. In particular, some implementations do want to use
the peer identity for access control decisions, though not on a per-user basis.

Dave Noveck suggested the new text. ]