Address IESG ballot COMMENT from Erik Kline
[ section 5.1.1 ]

* "When operation is complete" ... In addition to a grammar tweak, you
might repeat a few choice words from section 7.2 about the ability to
send multiple requests over a connection.

I was just thinking that it would be useful to add a sentence to
dispel any ideas that it's a "one TLS connection, one request" system.
7.2 explicitly talks about sending multiple requests per session:

"""
  Moreover, client implementations are free to transmit RPC requests
  for more than one RPC user using the same TLS session.
"""

[ I responded further:

Though not a typical use case, it's totally valid to establish a TLS
session for one or a handful of RPC operations. But I think what you
are saying is the document might be read to suggest a TLS(TCP) session
is valid for the exchange of only one RPC transaction, and of course
that is not at all the case. ]