Address DISCUSS and COMMENT positions from Benjamin Kaduk
[ These are all related to Section 5.2.1. The section needs to be
  rewritten to address all of these positions. ]

(5) Section 5.2.1 requires that:

  *  Implementations SHOULD indicate their trusted Certification
     Authorities (CAs).

Indicate to whom?

(6) The usage of RFC 6125 procedures in Section 5.2.1 seems counter
to its intent.  Specifically, we seem to be saying "the peer gave me
a cert, let me look through it to see if it has something I like",
but RFC 6125's intended procedure is "I know a list of names that I
expect to see at least one of in the cert; these rules tell me
whether the cert is valid for any such name".  It's not entirely
clear that it's appropriate for this document to specify how the
client has to order its list of names by type (per Section 6.1 of
RFC 6125's "The client constructs a list of acceptable reference
identifiers"), which the bit about "The following precedence
applies" seems to be doing.  To the extent that we give a
recommendation to use DNS-ID instead of CN-ID, and ipAddress SAN
instead of CN-ID, that's already covered by RFC 6125; it would be
okay for us to say "use DNS-ID or iPAddress SAN", though.  (Roman's
comment about "why not a normative MUST" for putting IP addresses in
the iPAddress SAN is related, and if we don't have a compelling
reason to allow the flexibility, we should limit to the specific
DNS-ID/iPAddress options without allowing CN-ID.)

(6.1) Note additionally that if wildcard certificates are to be
used, RFC 6125 requires the application protocol specification to
give details on how they are to be used.

(6.2) RFC 6125's procedures are (facially, at least) only valid for
TLS server authentication.  We also want to authenticate TLS
clients, so we should say whether we expect the same procedures to
be used, or what procedures should be used (even just as how it
differs from the RFC 6125 ones).  Of particular note is that, since
the server is not initiating the network connection, it is unlikely
to have a preconceived notion of what client identity to expect, and
is likely limited to attempting to extract something from the
certificate.  In this scenario a precedence list (as I complained
about being inconsistent with RFC 6125 above) would be appropriate.

(7) Section 5.2.1 uses the phrase "renegotiate the TLS session".
Renegotiation is not defined or allowed for TLS 1.3; generally one would
need to either remember the presented certificate and re-run the validation
process on it or shutdown the TLS connection and make a new one, though in
theory one could try to define a mechanism using post-handshake
authentication.  (I don't recommend the latter, though; it's not widely
implemented/used.)

[ COMMENT positions: ]

Section 5.2.1

  *  Certificate validation MUST include the verification rules as per
     [RFC5280].

It might be appropriate to reference RFC 6125 as well (though that places
some restrictions on how servers are named, claims to only apply to server
certificate validation, and requires specifying what type of name is to be
validated and how the input name to validation is obtained)

  *  Peer validation always includes a check on whether the locally
     configured expected DNS name or IP address of the server that is
     contacted matches its presented certificate.  DNS names and IP

In the RFC 6125 parlance these would be the "DNS-ID" and "iPAddress
subjectAltName", respectively, and such a reference would obviate the need
for most of the subsequent explanation.

  *  Implementations MAY allow the configuration of a set of additional
     properties of the certificate to check for a peer's authorization
     to communicate (e.g., a set of allowed values in
     subjectAltName:URI or a set of allowed X.509v3 Certificate
     Policies).

I could imagine an extendedKeyUsage being configured, too.

  trust model.  As a suggestion, at least the following parameters of
  the X.509 client certificate SHOULD be exposed:

What about non-extended keyUsage, and arbitrary X.509 extensions?

  *  Originating IP address

I don't think "originating IP address" is a parameter of the X.509 client
certificate.

  *  Certificate Fingerprint

Using what hash algorithm?

  *  Subject
  [...]
  *  all X.509v3 Subject Alternative Name

Per RFC 5280 (§ 4.1.2.6) the SAN *is* (part of) the Subject.