Address DISCUSS position from Benjamin Kaduk
(1) I don't think that the claim in Section 4.2 that "[b]oth RPC
and TLS have peer and user authentication" is correct, at least
given my understanding of those terms.  Using this document's
definition of RPC "peer authentication" as analogous to TLS server
authentication, the functionality that TLS calls "mutual
authentication" is more analogous to RPC client authentication,
though it is sometimes repurposed for use for user authentication,
with concommittant bad user experience.  This analogy does not seem
critical to the mechanisms of this document, so I believe we should
remove or modify it.

[ The comment was further clarified: the sentence in question
implies that TLS can provide user authentication, which it cannot. ]