Address DISCUSS position from Benjamin Kaduk
(8) Can we clarify the status of DNSSEC (or DANE) requirements?
Section 1 assumes that support for DNSSEC is already available in
an RPC implementation, but Section 7.1.1 says that "Clients [sic]
implementors can choose from" a list of things including DANE TLSA
records.  Why would we require DNSSEC support but not using the
records if they're present?

[ DNSSEC support is required, but we cannot require the use of
  TLSA records in environments where RPC-over-TLS is deployed. ]