Address DISCUSS position from Benjamin Kaduk
(4) I don't think it's particularly safe to suggest that non-
protected RPCs should be exchanged on the same 5-tuple that just
terminated a DTLS association, since neither DTLS nor UDP provide
in-order delivery, so there is ambiguity as to whether a datagram
should be interpreted as DTLS protected or not.  This is
particularly problematic in the face of the three different DTLS
record headers (DTLSPlaintext, DTLSCiphertext(full), and
DTLSCiphertext(minimal)) with something like 10 or 11 different
possible values for the first byte that might be in flight, with
limited "magic number" verification fields available.  I think I
need some input from the TSV ADs about what the options are,
though -- while a cooling-off period might be fine if an ephemeral
port is in use, it seems problematic for cases where fixed port
numbers are used for both source and destination.

[ Ben follows up with: ]

The main output of my conversation with the TSV ADs was that we
wanted to get your (and the WG's) thoughts about how RPC is actually
used, both for NFSv4 and elsewhere.  If there are not any
known/current deployments that would be affected, then we can
probably just document the risk and move on, but if there are
existing deployments that could trigger such a scenario, we'll need
to think through things more carefully.