Address COMMENT position from Benjamin Kaduk
Abstract

I can't quite put my finger on it, but something about "enables
encryption of in-transit Remote Procedure Call (RPC) transactions"
rubs me the wrong way.  I think I want to change what the "in
transit" binds to, since what we are doing is encrypting RPC
messages while they are in transit, as opposed to applying
protection to just the subset of messages that are "in transit".