Address COMMENT position from Benjamin Kaduk
Section 8

I think we should consider getting IANA to allocate OIDs for some X.509
extendedKeyUsage values that correspond to RPC-TLS client/server usage.
It's not necessarily something that we'd require to be present (but would be
helpful if a dedicated RPC-TLS PKI was to appear), but is useful to have
available for when it's needed.  This would presumably be under the
1.3.6.1.5.5.7.3 arc, and the referenced documents for existing allocations
usable as a template for what text to put in.