Address COMMENT position from Benjamin Kaduk
Section 4.2.1

  RPCSEC GSS can also perform per-request integrity or confidentiality
  protection.  When operating over a TLS session, these GSS services
  become redundant.  An RPC implementation capable of concurrently

At risk of being called a pedant, these services are only "largely
redundant" -- most TLS setups will be using standard asymmetric key
exchange that is vulnerable to a quantum computer, whereas a
kerberos-backed RPCSEC_GSS will be using symmetric crypto that is or
can be quantum-resistant for appropriate key sizes.  When a "record
now, decrypt later" attack is relevant the distinction can be
important to make (though for purely ephemeral traffic the
distinction collapses).