Address COMMENT position from Benjamin Kaduk
Section 5.1.2

Ben> I'm also curious about the addition of "RPC-over-DTLS support for that RPC
Ben> program on that port" -- why do we need to scope the discovery to a single
Ben> RPC program on that port for the UDP/DTLS case, when the TCP/TLS case just
Ben> says "RPC-over-TLS support for that connection"?

Chuck> RPC-over-(D)TLS discover must be done for each RPC program. That's true
Chuck> for TCP or UDP. You can't mix cleartext and TLS-protected traffic on
Chuck> a TCP connection with an established TLS session. However, with UDP,
Chuck> there is no connection that encases the DTLS session, and that allows
Chuck> mixture of cleartext and DTLS-protected datagrams to the same

Chuck> I'm open to suggestions (and editorial correction, of course).

Ben> Probably what I'm missing is how the TLS/TCP connection is bound to the
Ben> specific RPC program.  (I assume that it's just part of how RPC-over-TCP
Ben> works and I've forgotten that detail.)  Proceeding on that assumption, that
Ben> might suggest an editorial change to §5.1.1 to keep the TLS and DTLS
Ben> descriptions more parallel, perhaps "[...] it uses the mechanism described
Ben> in Section 4.1 to discover RPC-over-TLS support for the RPC program on that
Ben> connection."